Switched importAll to use global instead of this
I wrote importAll myself, and used this, and then I recalled that I had
seen a similar approach in https://github.com/arturadib/shelljs.

That project allows you to run the following:

    require('shelljs/global');

Which is what I think I want.

I think our implementations were sorta the same, but I was presuming
that "this" would be assigned to the global object whereas shelljs goes
right after the global object.

So.. I'll try that for now, too: this -> global.